move redis set/get to a common file
those function will be used by the "redis" and "wd" backend
