+++ pkg/keystore/redis.go
@@ -2,7 +2,6 @@ package keystore
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 
 	"github.com/go-redis/redis/v8"
@@ -184,24 +183,6 @@ func (s *RedisKeystore) GetDerivationPath(id uuid.UUID, address string) (Derivat
 
 func (s *RedisKeystore) MarkAddressAsUsed(id uuid.UUID, address string) error {
 	return keystoreMarkAddressAsUsed(s, id, address)
-}
-
-func set(c *redis.Client, key string, value interface{}) error {
-	p, err := json.Marshal(value)
-	if err != nil {
-		return err
-	}
-
-	return c.Set(context.Background(), key, string(p), 0).Err()
-}
-
-func get(c *redis.Client, key string, dest interface{}) error {
-	p, err := c.Get(context.Background(), key).Result()
-	if err != nil {
-		return err
-	}
-
-	return json.Unmarshal([]byte(p), dest)
 }
 
 func (s *RedisKeystore) GetAddressesPublicKeys(id uuid.UUID, derivations []DerivationPath) ([]string, error) {
